gralloc: Support Akebi96 with upstream ION driver
This adds Akebi96 board support with upstream ION driver.
To allocate framebuffer, this uses a CMA node which name has
'framebuffer'. So the board device tree has to define following
framebuffer node.

	reerved-memory {
		framebuffer@c2000000 {
			compatible = "shared-dma-pool";
			reg = <0 0xc2000000 0 0x02000000>;
			reusable;
		};
	};

The memory can't overlap with simple-fb memory because simple-fb
node must be defined as DMA node with 'no-map'.

Author: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>
Author: sato.shunji <sato.shunji@socionext.com>